# openapi.json
{
  "components": {
    "schemas": {
...
      "EventCreate": {
        "properties": {
          "chapter_id": {
            "title": "Chapter Id",
            "type": "string"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "feedback_enabled": {
            "default": true,
            "title": "Feedback Enabled",
            "type": "boolean"
          },
          "help_options": {
            "items": {
              "type": "string"
            },
            "title": "Help Options",
            "type": "array"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "locale": {
            "default": "nl",
            "enum": [
              "nl",
              "en"
            ],
            "title": "Locale",
            "type": "string"
          },
          "location": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "reminder_enabled": {
            "default": true,
            "title": "Reminder Enabled",
            "type": "boolean"
          },
          "source_options": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Options",
            "type": "array"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "topic": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          }
        },
        "required": [
          "name",
          "chapter_id",
          "location",
          "starts_at",
          "ends_at",
          "source_options"
        ],
        "title": "EventCreate",
        "type": "object"
      },
      "EventOut": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "attendee_count": {
            "title": "Attendee Count",
            "type": "integer"
          },
          "chapter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "chapter_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Name"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "feedback_enabled": {
            "title": "Feedback Enabled",
            "type": "boolean"
          },
          "help_options": {
            "items": {
              "type": "string"
            },
            "title": "Help Options",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "locale": {
            "enum": [
              "nl",
              "en"
            ],
            "title": "Locale",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reminder_enabled": {
            "title": "Reminder Enabled",
            "type": "boolean"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source_options": {
            "items": {
              "type": "string"
            },
            "title": "Source Options",
            "type": "array"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "topic",
          "location",
          "latitude",
          "longitude",
          "starts_at",
          "ends_at",
          "source_options",
          "help_options",
          "feedback_enabled",
          "reminder_enabled",
          "locale",
          "chapter_id",
          "chapter_name",
          "attendee_count",
          "archived"
        ],
        "title": "EventOut",
        "type": "object"
      },
      "EventStatsOut": {
        "description": "Organiser-only aggregate. Never includes individual signups.",
        "properties": {
          "by_help": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Help",
            "type": "object"
          },
          "by_source": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Source",
            "type": "object"
          },
          "total_attendees": {
            "title": "Total Attendees",
            "type": "integer"
          },
          "total_signups": {
            "title": "Total Signups",
            "type": "integer"
          }
        },
        "required": [
          "total_signups",
          "total_attendees",
          "by_source",
          "by_help"
        ],
        "title": "EventStatsOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SignupSummaryOut": {
        "description": "A single signup as seen on the organiser's details page.\nDeliberately minimal \u2014 name, headcount, help-choices. Never\nemail, source, or feedback-email status; those exist on the\nmodel but are private to the worker. ``id`` is exposed so the\norganiser can target an individual row for deletion (e.g.\ncleaning up an accidental sign-up they made themselves).",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "help_choices": {
            "items": {
              "type": "string"
            },
            "title": "Help Choices",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "party_size": {
            "title": "Party Size",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display_name",
          "party_size",
          "help_choices"
        ],
        "title": "SignupSummaryOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Opkomst",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "chapter_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response List Events Api V1 Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "create_event_api_v1_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/archived": {
      "get": {
        "operationId": "list_archived_events_api_v1_events_archived_get",
        "parameters": [
          {
            "in": "query",
            "name": "chapter_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response List Archived Events Api V1 Events Archived Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Archived Events",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/events/{event_id}": {
      "delete": {
        "description": "Hard-delete an archived event. Refuses if the event isn't\narchived first \u2014 accidentally hard-deleting a live event\nwith sign-ups would be a data-loss footgun. Cascades through\n``signups`` / ``email_dispatches`` / ``feedback_responses`` /\n``feedback_tokens`` via the FK ``ON DELETE CASCADE``s in the\nschema; the row + its dependents go with one DELETE.",
        "operationId": "delete_event_api_v1_events__event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Event",
        "tags": [
          "events"
        ]
      },
      "put": {
        "operationId": "update_event_api_v1_events__event_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}/archive": {
      "post": {
        "operationId": "archive_event_api_v1_events__event_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/events/{event_id}/restore": {
      "post": {
        "operationId": "restore_event_api_v1_events__event_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}/send-emails/{channel}": {
      "post": {
        "description": "Manually trigger the worker for one channel on a single\nevent. Re-use of the generic dispatcher means the reminder\n+ feedback \"send now\" buttons share one endpoint and one\nrate-limit budget.",
        "operationId": "send_emails_now_api_v1_events__event_id__send_emails__channel__post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "channel",
            "required": true,
            "schema": {
              "title": "Channel",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Send Emails Now Api V1 Events  Event Id  Send Emails  Channel  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Emails Now",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}/signups": {
      "get": {
        "description": "Per-signup list for the organiser details page. Returns\ndisplay_name + party_size + help_choices \u2014 never email,\nsource, or feedback-email status.",
        "operationId": "event_signups_api_v1_events__event_id__signups_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SignupSummaryOut"
                  },
                  "title": "Response Event Signups Api V1 Events  Event Id  Signups Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Signups",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/events/{event_id}/stats": {
      "get": {
        "operationId": "event_stats_endpoint_api_v1_events__event_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStatsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Stats Endpoint",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}